feat(b2b): Improve Company Admin View with share-admin components

+++ packages/b2b/src/workflows/employee/steps/create-employees.ts
@@ -7,7 +7,7 @@ import {
   ModuleEmployee,
 } from "@b2b/types";
 
-export const createEmployeesStep = createStep(
+export const createEmployeesStep = createStep<ModuleCreateEmployee, ModuleEmployee, string>(
   "create-employees",
   async (
     input: ModuleCreateEmployee,
@@ -33,7 +33,8 @@ export const createEmployeesStep = createStep(
 
     return new StepResponse(employee as unknown as ModuleEmployee, employee.id);
   },
-  async (employeeId: string, { container }) => {
+  async (employeeId: string | undefined, { container }) => {
+    if (!employeeId) return
     const companyModuleService =
       container.resolve<ICompanyModuleService>(COMPANY_MODULE);
     await companyModuleService.deleteEmployees([employeeId]);
